python-camera-system/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Pillow==10.0.0
dlib==19.24.2
requests==2.31.0
schedule==1.2.0
smtplib2==0.2.1
logging==0.4.9.6
threading2==0.1.2
